Additional program build support.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@2644 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ libecc/usrinclude/sparc/linux/sys/types.h
@@ -172,7 +172,7 @@ typedef	int32_t		off_t;		/* file offset */
 #define	off_t		int32_t
 #endif
 
-typedef	uint16_t	pid_t;		/* process id */
+typedef	int32_t	        pid_t;		/* process id */
 typedef int32_t		lwpid_t;	/* LWP id */
 typedef uint64_t	rlim_t;		/* resource limit */
 typedef	int32_t		segsz_t;	/* segment size */
